Add red horizontal line in place of deleted cells
For every deleted cell, put a red horizontal line to signify the user that the kernel hasn't known about this change. 
Once the user undelete the cell or the kernel knows about the change, the red horizontal line disappear.
The user can also undelete a cell by clicking on the corresponding red line.
The old behavior of undelete cell is maintained.