Rewrite Depviewer
This is the primary commit in here, it's very hard to stage this into chunks. I almost entirely rewrote the Depviewer into a RequireJS module, removed a lot of unnecessary code and changed the modeling approach from Dagre-d3 to d3-graphviz. This is a major rewrite so the codebase doesn't even look very similar.