Disallow dynamic unscoped references
Each cell now needs to explicitly define any reference. Thus if there is a dynamic reference to a variable, the user should add a "var = var" line to the beginning of the cell.

Example:

[1] df.query('column = @column_value')

should instead be

[1] column_value = column_value
    df.query('column = @column_value')

The column_value will then be properly linked to the corresponding cell (column_value$ffffff)