Fixes #12
Formats the unnamed tuples in a uniform manner, namely as [ce111d][0].

Note that there are still problems with this approach as a single output is stored in a tuple when there is a library, meaning its access does not align with the posted reference. Also, if the cell is tagged, this reference is not updated as it is set at the time the output is created, but I don't like just abbreviating to [0], [1] since this conflicts with the standard notebook output notation for different cell outputs.